test(observers): re-sync spec structure for target-accessors
